% Auto-generated: do not edit by hand
\name{''PanelGroup}

\alias{''PanelGroup}

\title{PanelGroup component}

\description{

}

\usage{
''PanelGroup(children=NULL, id=NULL, autoSaveId=NULL, className=NULL,
direction=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{autoSaveId}{Character. Unique id used to aut-save group arragement via local storage.}

\item{className}{Character. class name for the panel group}

\item{direction}{Character. Direction of the panel group}

\item{style}{Named list. style for the panel group}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

