% Auto-generated: do not edit by hand
\name{''PanelResizeHandle}

\alias{''PanelResizeHandle}

\title{PanelResizeHandle component}

\description{

}

\usage{
''PanelResizeHandle(children=NULL, id=NULL, className=NULL, disable=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. class name for the panel group}

\item{disable}{Logical. Disable drag handle}

\item{style}{Named list. style for the panel group}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

