% Auto-generated: do not edit by hand
\name{''Panel}

\alias{''Panel}

\title{Panel component}

\description{

}

\usage{
''Panel(children=NULL, id=NULL, className=NULL,
collapsedSizePercentage=NULL, collapsedSizePixels=NULL,
collapsible=NULL, defaultSizePercentage=NULL,
defaultSizePixels=NULL, maxSizePercentage=NULL,
maxSizePixels=NULL, minSizePercentage=NULL,
minSizePixels=NULL, order=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. class name for the panel group}

\item{collapsedSizePercentage}{Numeric. Panel should collapse to this size (in percentage)}

\item{collapsedSizePixels}{Numeric. Panel should collapse to this size (in pixesl)}

\item{collapsible}{Logical. Panel should collapse when resized beyond its minSize}

\item{defaultSizePercentage}{Numeric. Initial size of panel (in percentage)}

\item{defaultSizePixels}{Numeric. Initial size of panel (in pixels)}

\item{maxSizePercentage}{Numeric. Maximum size of panel (in percentage)}

\item{maxSizePixels}{Numeric. Maximum size of panel (in pixels)}

\item{minSizePercentage}{Numeric. Minimum size of panel (in percentage)}

\item{minSizePixels}{Numeric. Minimum size of panel (in pixels)}

\item{order}{Numeric. Order of panel within group; required for groups with conditionally rendered panels}

\item{style}{Named list. style for the panel group}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

